Dynamic allocation cleanup of dynarray
Some things just shouldn't be put on the heap. This commit aims to fix this quote on quote issue. I discovered this little project on reddit, and decided to give it a look. The first file I opened was dynarray.c, and I got a little spooked for a second. I got rid of some unnecessary heap allocs, while trying to be as close to the original style of the codebase as possible