Use references only in the right situation
I've come across a lot of function signatures in this codebase that ask for a reference to something, without the need of mutability. I tried fixing these, but the codebase is a little hard to follow, especially when it comes to `board.c`. In the end I hope this commit communicates what my intention was.